requirements.txt
chardet==5.2.0
einops==0.8.0
g2p_en==2.1.0
h5py==3.11.0
librosa==0.10.1
matplotlib==3.8.2
nltk==3.8.1
numba==0.60.0
numpy==1.24.1
packaging==24.2
pandas==2.2.3
praat-parselmouth==0.4.5
pycwt==0.4.0b0
pyloudnorm==0.1.1
PyYAML==6.0.1
Resemblyzer==0.1.4
scikit_learn==1.5.2
scipy
six==1.16.0
tensorflow
tensorboard
textgrid==1.6.1
tqdm==4.66.5
webrtcvad==2.0.10
scikit-image==0.25.1
torchtyn==1.0.6
